feat: library improvements and fixes
- Import from URL with read timeout (30s connect, 120s per chunk)
- Live import progress: size updates in DB after each chunk
- Auto-refresh library page while items importing
- Edit/rename library items via PATCH endpoint
- Library picker updates disk size and format on selection
- Size field uses Math.max to keep larger user-specified size
- BigInteger for size_bytes (supports files >2GB)
- Fixed duplicate ISO badge, shows "starting download..." at 0 bytes
- Delete disabled during import

+++ src/backend/app/models/library.py
@@ -1,7 +1,7 @@
 import datetime
 import uuid
 
-from sqlalchemy import DateTime, ForeignKey, Integer, String, func
+from sqlalchemy import BigInteger, DateTime, ForeignKey, Integer, String, func
 from sqlalchemy.dialects.postgresql import JSONB, UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
@@ -30,7 +30,7 @@ class LibraryItem(Base):
     description: Mapped[str | None] = mapped_column(String(1000))
     type: Mapped[str] = mapped_column(String(20))
     format: Mapped[str] = mapped_column(String(10))
-    size_bytes: Mapped[int] = mapped_column(Integer, default=0)
+    size_bytes: Mapped[int] = mapped_column(BigInteger, default=0)
     s3_key: Mapped[str | None] = mapped_column(String(500))
     checksum_sha256: Mapped[str | None] = mapped_column(String(64))
     os_variant: Mapped[str | None] = mapped_column(String(50))
